+++ .github/workflows/test.yaml
@@ -32,4 +32,4 @@ jobs:
       - name: Upload Coverage
         env:
           COVERALLS_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: goveralls -coverprofile=covprofile -service=github
+        run: goveralls -ignore=cmd/ -coverprofile=covprofile -service=github
