+++ README.md
@@ -4,14 +4,13 @@
 [![CI Build](https://github.com/mrwormhole/emailer/actions/workflows/test.yaml/badge.svg)](https://github.com/mrwormhole/emailer/actions/workflows/test.yaml)
 [![GoDoc](https://godoc.org/github.com/mrwormhole/emailer?status.svg)](https://godoc.org/github.com/mrwormhole/emailer)
 [![Report Card](https://goreportcard.com/badge/github.com/mrwormhole/emailer)](https://goreportcard.com/report/github.com/mrwormhole/emailer)
-[![License](https://img.shields.io/github/license/mrwormhole/emailer)](https://github.com/mrwormhole/emailer/blob/master/LICENSE)
-
-tiny email server for managed email providers
+[![License](https://img.shields.io/github/license/mrwormhole/emailer)](https://github.com/mrwormhole/emailer/blob/main/LICENSE)
+[![Coverage Status](https://coveralls.io/repos/github/mrwormhole/emailer/badge.svg?branch=main)](https://coveralls.io/github/mrwormhole/emailer?branch=main)
 
 ### Purpose
 
-packaging SMTP(send only) APIs to a choosable option is the goal here, then to have a tiny image that picks 1 provider and serves
-as HTTP server so other microservices that belong to you can make email sending for multiple clients of yours.
+Packaging SMTP(send only) APIs to a choosable option is the goal here, then to have a tiny server that picks 1 provider and serves
+as HTTP endpoint so the other microservices that belong to you can make email requests.
 
 ### Supported providers
